# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suxen control-plane API",
    "version": "dev",
    "description": "Administrative API for repositories, access control, retention, provenance, and operations. OCI Distribution endpoints under /v2 and /repository/{name}/v2 use the OCI error envelope and are intentionally outside this control-plane contract."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current suxen server"
    }
  ],
  "security": [
    {},
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    },
    {
      "oidcSession": []
    }
  ],
...
  "paths": {
...
    "/api/v1/download-gate-defaults": {
      "get": {
        "operationId": "getDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:read",
        "summary": "Get the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "responses": {
          "200": {
            "description": "Instance-wide download-gate default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "setDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:write",
        "summary": "Set the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadGateDefaultsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instance-wide download-gate default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "201": {
            "description": "Instance-wide download-gate default created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DownloadGate"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:write",
        "summary": "Delete the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "responses": {
          "204": {
            "description": "Instance-wide download-gate default deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/users/{name}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Username"
        }
      ],
      "get": {
        "operationId": "getUser",
        "x-suxen-required-privilege": "admin:users:read",
        "summary": "Get a local user and roles",
        "tags": [
          "Access"
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "x-suxen-required-privilege": "admin:users:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Update a local user",
        "tags": [
          "Access"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "201": {
            "description": "User created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "x-suxen-required-privilege": "admin:users:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Delete a local user",
        "tags": [
          "Access"
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/{name}/roles": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Username"
        }
      ],
      "get": {
        "operationId": "getUserRoles",
        "x-suxen-required-privilege": "admin:users:read",
        "summary": "Get roles assigned to a user",
        "tags": [
          "Access"
        ],
        "responses": {
          "200": {
            "description": "User roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoles"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "setUserRoles",
        "x-suxen-required-privilege": "admin:users:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Replace roles assigned to a user",
        "tags": [
          "Access"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role assignments",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRoles"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/gc": {
      "post": {
        "operationId": "runGarbageCollection",
        "x-suxen-required-privilege": "admin:gc:run",
        "summary": "Run blob garbage collection",
        "description": "Invokes the builtin gc job. Task history remains under /api/v1/tasks.",
        "tags": [
          "Operations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DryRun"
          },
          {
            "$ref": "#/components/parameters/GracePeriod"
          }
        ],
        "responses": {
          "200": {
            "description": "Garbage collection result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarbageCollectionResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Local username and password."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token or OIDC JWT",
        "description": "Suxen API token or a configured OIDC bearer token. API-token scopes use repository:<name|*>:<read|write|delete|annotate|manage> and admin:<resource|*>:<action|*>. An empty scope list does not further restrict the user's roles.",
        "x-suxen-scope-syntax": [
          "repository:<name|*>:<read|write|delete|annotate|manage>",
          "admin:<resource|*>:<action|*>"
        ]
      },
      "oidcSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suxen_oidc_token",
        "description": "HTTP-only browser session established by the configured OIDC login flow."
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical path of the created resource.",
        "schema": {"type": "string"}
      }
    },
    "parameters": {
...
      "Username": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "Username",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._@+-]*$"
        }
      },
...
      "DryRun": {
        "name": "dryRun",
        "in": "query",
        "required": false,
        "description": "Report changes without deleting",
        "schema": {
          "type": "boolean",
          "default": true
        }
      },
...
      "Force": {
        "name": "force",
        "in": "query",
        "required": false,
        "description": "Transfer a managed resource out of declarative provisioning before applying the mutation",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "GracePeriod": {
        "name": "grace",
        "in": "query",
        "required": false,
        "description": "Minimum unreferenced blob age as a Go duration",
        "schema": {
          "type": "string",
          "default": "24h",
          "example": "24h"
        }
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. Account validation codes include invalid_username and password_required; the reserved repository name default returns reserved_repository_name.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required. Control-plane routes answer 401 without a WWW-Authenticate challenge so browsers do not raise the native Basic-auth popup; clients authenticate with a bearer token or the cookie session and treat 401 as a prompt to sign in.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient privilege",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists, is declaratively managed, or is still referenced",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 7807 problem details with a stable suxen error code.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:suxen:problem:not_found"
          },
          "title": {
            "type": "string",
            "example": "Not found"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9_]+$",
            "example": "not_found"
          }
        }
      },
...
      "DownloadGateRequest": {
        "type": "object",
        "required": ["criteria"],
        "properties": {
          "criteria": {"$ref": "#/components/schemas/Criteria"},
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "inheritGlobal": {
            "type": "boolean",
            "default": true,
            "description": "When true, the instance-wide default gate's criteria AND-extend this repository's."
          }
        }
      },
...
      "DownloadGateDefaultsUpdateRequest": {
        "type": "object",
        "description": "Replaces the instance-wide gate. criteria must be a non-null array; [] clears criteria. Missing or null criteria are rejected without changing policy or ownership.",
        "additionalProperties": false,
        "required": ["criteria"],
        "properties": {
          "criteria": {"$ref": "#/components/schemas/CriteriaRequest"},
          "enabled": {"type": "boolean", "default": true}
        }
      },
      "DownloadGate": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DownloadGateRequest"
          },
          {
            "type": "object",
            "required": [
              "repository",
              "managed",
              "updatedAt"
            ],
            "properties": {
              "repository": {
                "type": "string"
              },
              "managed": {
                "type": "boolean",
                "readOnly": true
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Timestamp"
              }
            }
          }
        ]
      },
...
      "User": {
        "type": "object",
        "required": [
          "username",
          "admin",
          "managed",
          "createdAt"
        ],
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 255
          },
          "admin": {
            "type": "boolean"
          },
          "managed": {
            "type": "boolean",
            "readOnly": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "description": "Updates only supplied fields. An omitted admin field preserves the current value within the update transaction. An empty or omitted password preserves the current password.",
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "writeOnly": true
          },
          "admin": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UserRoles": {
        "type": "object",
        "required": [
          "username",
          "roles"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserRolesRequest": {
        "type": "object",
        "description": "Replaces role assignments. roles must be a non-null array; [] clears assignments. Missing or null roles are rejected without changing assignments.",
        "additionalProperties": false,
        "required": [
          "roles"
        ],
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Criteria": {
        "type": "array",
        "description": "Ordered, conjunctive policy predicates.",
        "minItems": 1,
        "items": {
          "$ref": "#/components/schemas/Predicate"
        }
      },
      "PredicateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "op"],
        "properties": {
          "path": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*$"
          },
          "op": {
            "type": "string",
            "enum": [
              "=", "!=", "<", "<=", ">", ">=", "before", "after",
              "matches", "contains", "in", "not-in", "exists", "absent"
            ]
          },
          "value": {
            "description": "Operator literal. Omit for exists and absent."
          }
        }
      },
      "CriteriaRequest": {
        "type": "array",
        "description": "Ordered, conjunctive policy predicates.",
        "minItems": 1,
        "items": {"$ref": "#/components/schemas/PredicateRequest"}
      },
...
      "GarbageCollectionResult": {
        "type": "object",
        "required": [
          "dryRun",
          "scanned",
          "referenced",
          "deleted",
          "reclaimedBytes",
          "gracePeriod",
          "staleUploads",
          "staleUploadsDeleted",
          "staleStagingFiles",
          "staleStagingFilesDeleted"
        ],
        "properties": {
          "dryRun": {
            "type": "boolean"
          },
          "blobStore": {
            "type": "string",
            "description": "Set when collection was scoped to one blob store; absent for a global run."
          },
          "scanned": {
            "type": "integer",
            "minimum": 0
          },
          "referenced": {
            "type": "integer",
            "minimum": 0
          },
          "deleted": {
            "type": "integer",
            "minimum": 0
          },
          "reclaimedBytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "wouldDelete": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gracePeriod": {
            "type": "string"
          },
          "staleUploads": {
            "type": "integer",
            "minimum": 0
          },
          "staleUploadsDeleted": {
            "type": "integer",
            "minimum": 0
          },
          "staleStagingFiles": {
            "type": "integer",
            "minimum": 0,
            "description": "Abandoned generic staging files found on this replica by global GC."
          },
          "staleStagingFilesDeleted": {
            "type": "integer",
            "minimum": 0,
            "description": "Abandoned generic staging files removed on this replica; zero for dry runs."
          }
        }
      },
...
    }
  }
}